+++ packages/configuration-all/package.json
@@ -9,7 +9,6 @@
     "@hint/formatter-json": "^3.1.0",
     "@hint/formatter-stylish": "^3.1.0",
     "@hint/formatter-summary": "^3.0.3",
-    "hint": "^5.1.0",
     "@hint/hint-amp-validator": "^2.9.0",
     "@hint/hint-apple-touch-icons": "^3.1.0",
     "@hint/hint-axe": "^4.0.0",
@@ -60,8 +59,7 @@
     "@hint/parser-manifest": "^2.2.1",
     "@hint/parser-package-json": "^1.1.3",
     "@hint/parser-typescript-config": "^2.3.0",
-    "@hint/parser-webpack-config": "^2.1.3",
-    "@hint/utils": "^3.0.0"
+    "@hint/parser-webpack-config": "^2.1.3"
   },
   "description": "webhint's all configuration that includes all the published packages",
   "engines": {
