+++ release/tasks/calculate-new-versions.ts
@@ -1,7 +1,8 @@
-import { Context, Bump, Tag } from '../@types/custom';
+import { Context, Bump, Tag, Package } from '../@types/custom';
+import * as semver from 'semver';
 
+import { debug } from '../lib/utils';
 import updateDependencies from '../lib/update-dependencies';
-import calculatePackageNewVersion from '../lib/calculate-package-new-version';
 
 /**
  *
@@ -20,6 +21,25 @@ const getBumpTypeForTag = (tag: Tag): Bump => {
         case 'Breaking': return Bump.major;
         default: return Bump.none;
     }
+};
+
+const calculatePackageNewVersion = (pkg: Package, bump: Bump): string => {
+
+    if (pkg.ignore) {
+        return pkg.content.version;
+    }
+
+    if (!pkg.publishedVersion) {
+        debug(`${pkg.name} will be published with initial version ${pkg.content.version}`);
+
+        return pkg.content.version;
+    }
+
+    const newVersion = semver.inc(pkg.oldVersion, Bump[bump] as semver.ReleaseType)!;
+
+    debug(`Bumping ${pkg.name} from ${pkg.oldVersion} to ${newVersion}`);
+
+    return newVersion;
 };
 
 /**
